move clients to its own directory
